requirements.txt
Bootstrap_Flask==2.3.3
Flask~=3.1.1
Flask-Bootstrap~=3.3.7.1
pandas~=2.3.1
requests~=2.32.4
gspread~=6.2.1
oauth2client~=4.1.3
python-dotenv~=1.1.1